Etap_I/API1/app/requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
requests
PyJWKClient
pymysql